requirements.txt
# Telegram Bot Framework
aiogram==3.7.0

# Web Framework for webhooks
fastapi==0.115.0
uvicorn==0.30.1

# Database
motor==3.5.1
pymongo==4.8.0

# Payment Gateways
razorpay==1.4.2
Pillow==10.4.0

# HTTP Requests
httpx==0.27.0
requests==2.32.3
